refactor(player): 实现 DesktopAudioVisualManager 并调整 VideoInfo序列化
- 实现 DesktopAudioVisualManager 类，提供音频视觉控制的基本功能
- 修改 VideoInfo 中 DescV2 类的 bizId 字段类型，从 Int 改为 Long

+++ biz/biliplayer/src/commonMain/kotlin/top/sacz/bili/biz/biliplayer/entity/VideoInfo.kt
@@ -202,7 +202,7 @@ data class VideoInfo(
     @Serializable
     data class DescV2(
         @SerialName("biz_id")
-        val bizId: Int = 0, // 示例值：0
+        val bizId: Long = 0, // 示例值：0
         @SerialName("raw_text")
         val rawText: String = "", // 示例值："【CB想说的】看完拜年祭之后最爱的一个节目！..."
         @SerialName("type")
